Make Widget an interface
